+++ src/mesido/financial_mixin.py
@@ -125,17 +125,6 @@ class FinancialMixin(
                 )
                 nominal_variable_operational = nominal_fixed_operational
                 nominal_investment = nominal_fixed_operational
-            elif asset_name in [*self.energy_system_components.get("low_temperature_ates", [])]:
-                nominal_fixed_operational = self.variable_nominal(
-                    f"{asset_name}.Heat_low_temperature_ates"
-                )
-                nominal_fixed_operational = (
-                    nominal_fixed_operational
-                    if isinstance(nominal_fixed_operational, float)
-                    else max(nominal_fixed_operational.values)
-                )
-                nominal_variable_operational = nominal_fixed_operational
-                nominal_investment = nominal_fixed_operational
             elif asset_name in [*self.energy_system_components.get("heat_demand", [])]:
                 nominal_fixed_operational = (
                     bounds[f"{asset_name}.Heat_demand"][1]
@@ -450,7 +439,6 @@ class FinancialMixin(
             *self.energy_system_components.get("heat_demand", []),
             *self.energy_system_components.get("cold_demand", []),
             *self.energy_system_components.get("ates", []),
-            *self.energy_system_components.get("low_temperature_ates", []),
             *self.energy_system_components.get("heat_buffer", []),
             *self.energy_system_components.get("heat_pipe", []),
             *self.energy_system_components.get("heat_exchanger", []),
@@ -478,7 +466,6 @@ class FinancialMixin(
                 *self.energy_system_components.get("cold_demand", []),
                 *self.energy_system_components.get("heat_pipe", []),
                 *self.energy_system_components.get("ates", []),
-                *self.energy_system_components.get("low_temperature_ates", []),
                 *self.energy_system_components.get("heat_buffer", []),
                 *self.energy_system_components.get("heat_exchanger", []),
                 *self.energy_system_components.get("heat_pump", []),
@@ -934,7 +921,6 @@ class FinancialMixin(
 
         for asset in [
             *self.energy_system_components.get("ates", []),
-            *self.energy_system_components.get("low_temperature_ates", []),
             *self.energy_system_components.get("heat_buffer", []),
         ]:
             heat_charge = self.__state_vector_scaled(f"{asset}.Heat_flow_charging", ensemble_member)
@@ -1068,6 +1054,22 @@ class FinancialMixin(
                 / denominator
             )
             sum_ += ca.sum1(price_profile.values[1:] * pump_power[1:] * timesteps_hr / eff)
+
+            if (len(self.get_electricity_carriers().keys()) > 0) and s in [
+                *self.energy_system_components.get("heat_source_elec", []),
+                *self.energy_system_components.get("elec_heat_source_elec", []),
+                *self.energy_system_components.get("air_water_heat_pump_elec", []),
+                *self.energy_system_components.get("heat_pump_elec", []),
+            ]:
+                sum_ += (
+                    ca.sum1(price_profile.values[1:] * nominator_vector[1:] * timesteps_hr)
+                    / denominator
+                )
+                if variable_operational_cost_coefficient > 0.0:
+                    logger.warning(
+                        f"Variable operational cost for {s} is derived from both the variable "
+                        "operational cost coefficient and the electricity carrier cost."
+                    )
 
             constraints.append(((variable_operational_cost - sum_) / nominal, 0.0, 0.0))
 
@@ -1307,7 +1309,6 @@ class FinancialMixin(
                 *self.energy_system_components.get("heat_source", []),
                 *self.energy_system_components.get("heat_pipe", []),
                 *self.energy_system_components.get("ates", []),
-                *self.energy_system_components.get("low_temperature_ates", []),
                 *self.energy_system_components.get("heat_buffer", []),
                 *self.energy_system_components.get("heat_exchanger", []),
                 *self.energy_system_components.get("heat_pump", []),
@@ -1411,7 +1412,6 @@ class FinancialMixin(
                 *self.energy_system_components.get("heat_source", []),
                 *self.energy_system_components.get("heat_pipe", []),
                 *self.energy_system_components.get("ates", []),
-                *self.energy_system_components.get("low_temperature_ates", []),
                 *self.energy_system_components.get("heat_buffer", []),
                 *self.energy_system_components.get("heat_exchanger", []),
                 *self.energy_system_components.get("heat_pump", []),
@@ -1517,7 +1517,6 @@ class FinancialMixin(
         asset_categories = [
             "heat_source",
             "ates",
-            "low_temperature_ates",
             "heat_buffer",
             "heat_pipe",
             "heat_exchanger",